# openapi.yaml
openapi: 3.0.3
info:
  title: Excalidraw FULL API
  description: API for the visual workspace platform
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/auth/signup:
    post:
      operationId: signup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 6 }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  session: { $ref: '#/components/schemas/Session' }
        '429': { description: Rate limited }
  /api/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: '#/components/schemas/User' }
                  session: { $ref: '#/components/schemas/Session' }
        '401': { description: Invalid credentials }
...
  /api/auth/me:
    get:
      operationId: getMe
      security: [ { cookieAuth: [] } ]
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: { $ref: '#/components/schemas/User' }
        '401': { description: Not authenticated }
  /api/teams:
    get:
      operationId: listTeams
      security: [ { cookieAuth: [] } ]
      responses:
        '200':
          description: List of teams
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Team' } }
    post:
      operationId: createTeam
      security: [ { cookieAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTeamRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Team' }
  /api/teams/{teamID}/members:
    get:
      operationId: listTeamMembers
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: teamID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Members
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TeamMember' } }
  /api/teams/{teamID}/invites:
    get:
      operationId: listTeamInvites
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: teamID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Invites
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/TeamInvite' } }
    post:
      operationId: createTeamInvite
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: teamID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInviteRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  invite: { $ref: '#/components/schemas/TeamInvite' }
                  token: { type: string }
  /api/invites/accept:
    post:
      operationId: acceptInvite
      security: [ { cookieAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token: { type: string }
      responses:
        '200':
          description: Membership created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/TeamMember' }
  /api/drawings:
    get:
      operationId: listDrawings
      security: [ { cookieAuth: [] } ]
      parameters:
        - { name: team_id, in: query, schema: { type: string } }
        - { name: folder_id, in: query, schema: { type: string } }
      responses:
        '200':
          description: Drawings
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Drawing' } }
    post:
      operationId: createDrawing
      security: [ { cookieAuth: [] } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateDrawingRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id: { type: string }
  /api/drawings/{drawingID}:
    get:
      operationId: getDrawing
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Drawing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Drawing' }
        '404': { description: Not found }
    patch:
      operationId: updateDrawing
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateDrawingRequest' }
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated_at: { type: string, format: date-time }
        '404': { description: Not found }
    delete:
      operationId: archiveDrawing
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '204': { description: Archived }
  /api/drawings/{drawingID}/revisions:
    get:
      operationId: listRevisions
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Revisions
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/DrawingRevision' } }
    post:
      operationId: createRevision
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [change_summary, snapshot]
              properties:
                change_summary: { type: string, maxLength: 240 }
                snapshot: { type: string }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DrawingRevision' }
  /api/drawings/{drawingID}/permissions:
    get:
      operationId: listPermissions
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Grants
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/PermissionGrant' } }
    post:
      operationId: createPermission
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreatePermissionGrantRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PermissionGrant' }
  /api/drawings/{drawingID}/share-links:
    get:
      operationId: listShareLinks
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Links
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/ShareLink' } }
    post:
      operationId: createShareLink
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateShareLinkRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  share_link: { $ref: '#/components/schemas/ShareLink' }
                  token: { type: string }
  /api/drawings/{drawingID}/assets:
    get:
      operationId: listAssets
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Assets
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Asset' } }
    post:
      operationId: createAsset
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateAssetRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Asset' }
  /api/drawings/{drawingID}/embeds:
    get:
      operationId: listEmbeds
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Embeds
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Embed' } }
    post:
      operationId: createEmbed
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateEmbedRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Embed' }
  /api/drawings/{drawingID}/links:
    get:
      operationId: listLinks
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Links
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/LinkReference' } }
    post:
      operationId: createLink
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: drawingID, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateLinkRequest' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/LinkReference' }
  /api/shared/{token}:
    get:
      operationId: getSharedResource
      parameters: [ { name: token, in: path, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Shared drawing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SharedResource' }
        '404': { description: Invalid or expired token }
  /api/search:
    get:
      operationId: searchDrawings
      security: [ { cookieAuth: [] } ]
      parameters: [ { name: q, in: query, required: true, schema: { type: string } } ]
      responses:
        '200':
          description: Results
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Drawing' } }
  /api/templates:
    get:
      operationId: listTemplates
      security: [ { cookieAuth: [] } ]
      responses:
        '200':
          description: Templates
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/Template' } }
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: excalidraw_session
  schemas:
...
    User:
      type: object
      required: [id, name, username, email]
      properties:
        id: { type: string }
        name: { type: string }
        username: { type: string }
        email: { type: string }
        avatar_url: { type: string, nullable: true }
        locale: { type: string }
        timezone: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    Session:
      type: object
      required: [id, user_id, token, expires_at]
      properties:
        id: { type: string }
        user_id: { type: string }
        token: { type: string }
        expires_at: { type: string, format: date-time }
        created_at: { type: string, format: date-time }
    Team:
      type: object
      required: [id, name, slug, created_by]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
        created_by: { type: string }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
...
    CreateTeamRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string, nullable: true }
        avatar_url: { type: string, nullable: true }
...